Added secret_name parameter to get_secret, moved util function to utils.py and pytest tests for util functions to test_utils.py
